Optimize performance when using files whitelist.
The original file whitelisting process works by using a
FileWhitelistScoper class that maintains a list of all the whitelisted
files. Every time a file is to be prefixed, the list is scanned and
if it contains the file, then it's not prefixed. This is inefficient
when the list of whitelisted files is big.

This commit changes the process to use a different approach.
First, the whitelisted files are excluded from the list of files
generated by the finders (and whose contents are loaded in memory).
Then both lists are passed to the AddPrefixCommand: the matched files
are processed as usual, the whitelisted ones are simply copied verbatim
using fileSystem->copy.

The logger is modified too so that now at the end of the process
two separate counts are displayed, one for the prefixed files and
another one for the copied (whitelisted) files.